js : c(normal) : c(void ) modified delete_node

+++ c/normal/binary_search_tree.c
@@ -65,16 +65,16 @@ Node_ptr delete_node(Node_ptr root, int value)
 {
   if (root == NULL)
     return root;
-  if (value < root->value)
-    root->left = delete_node(root->left, value);
-  if (value > root->value)
-    root->right = delete_node(root->right, value);
+  root->left = value < root->value ? delete_node(root->left, value) : root->left;
+  root->right = value > root->value ? delete_node(root->right, value) : root->right;
   if (value == root->value)
   {
-    if (root->left == NULL)
-      return root->right;
-    if (root->right == NULL)
-      return root->left;
+    if (root->left == NULL || root->right == NULL)
+    {
+      Node_ptr temp = root->left ? root->left : root->right;
+      free(root);
+      return temp;
+    }
     Node_ptr minOfRight = get_min_of_right(root->right);
     root->value = minOfRight->value;
     root->right = delete_node(root->right, minOfRight->value);
